ShortCommit

+++ src/CharacterFrequencyCounterTest.java
@@ -1,7 +1,5 @@
 import static org.junit.Assert.assertEquals;
 
-import javax.xml.crypto.dsig.spec.C14NMethodParameterSpec;
-
 import org.junit.jupiter.api.Test;
 
 public class CharacterFrequencyCounterTest {
@@ -31,7 +29,7 @@ public class CharacterFrequencyCounterTest {
             //Arrange
             CharacterFrequencyCounter counter = new CharacterFrequencyCounter("$%^&");
             //Act
-            int actual = counter.getRelativePercentage('1');
+            double actual = counter.getRelativePercentage('1');
             //Assert
             assertEquals(0.0, actual, .001);
     }
